TablingReservationRequest Rtes + Unavailable Opts
Add functionality for student orgs to fill out the EMS reservation directly on our website (hopefully we can figure out a script so that we plug in our options for the user to make the reservation on the EMS website/do the role of the EMS website).
I also added functionality to display the location of all unavailable locations and times of reserved places for tabling (so users cannot pick those spots).
Finally, for the form, I allowed users to request their most recent submission(s) and use it for a future submission